source/pyproject.toml
[tool.poetry]
name = "esdp"
version = "0.1.0"
description = ""
authors = ["Tamilia <apfellfcnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
factory-boy = "^3.2.1"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
pycodestyle = "^2.10.0"
django-constance = "^2.9.1"
django-redis = "^5.2.0"
aiogram = "^2.25.1"
requests = "^2.28.2"
pillow = "^9.4.0"
selenium = "^4.8.3"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"